+++ permission_handler/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 6.0.0
 
+This release includes the following **breaking changes**:
 * Migrated to null safety 
+* Removed PermissionStatus.undetermined, PermissionStatus.denied will be used instead.
 
 ## 5.1.0+2